added network generators
added network generators, tests, html improvements. Broken down the util module into smaller modular scripts.
