+++ main.py
@@ -1,8 +1,6 @@
 import uvicorn
 from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
-from fastapi.responses import RedirectResponse
-from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from routers.endpoints import endpoints
 from routers.views import views
@@ -11,11 +9,6 @@ app = FastAPI()
 app.include_router(endpoints)
 app.include_router(views)
 app.mount("/static", StaticFiles(directory="static"), name="static")
-
-
-@app.exception_handler(StarletteHTTPException)
-async def custom_http_exception_handler(request, exc):
-    return RedirectResponse("/404")
 
 
 if __name__ == "__main__":
