+++ templates/index.html
@@ -8,7 +8,7 @@
     <meta name="description"
           content="StackDeck is a blazing fast api "/>
     <meta property="og:title" content="StackDeck Blazing fast api for searching tech stacks.">
-    <meta property="og:description" content="Browse trougth the stacks of diferent companies compare and analyze tech popularity.">
+    <meta property="og:description" content="Browse througth the stacks of different companies compare and analyse tech popularity.">
     <meta property="og:type" content="article">
     <meta property="og:url" content="https://stackdeck.dev">
     <meta property="og:image" content="https://stackdeck.dev/static/images/og_image.png"/>
